+++ packages/common/src/Adaptor.js
@@ -80,7 +80,7 @@ export function fn(func) {
 }
 
 /**
- * A custom operation that will only execute the function if the condition returns true
+ * Execute a function only when the condition returns true
  * @public
  * @function
  * @example