+++ packages/pdfshift/src/Adaptor.js
@@ -1,4 +1,4 @@
-import { expandReferences, encode } from '@openfn/language-common/util';
+import { expandReferences } from '@openfn/language-common/util';
 import * as util from './Utils';
 
 /**
@@ -70,7 +70,7 @@ import * as util from './Utils';
  *  filename: 'example.pdf',
  * });
  * @returns {Operation}
- * @state {HttpState}
+ * @state data - The parsed response body when `filename` is provided, otherwise a base64-encoded string.
  */
 export function generatePDF(htmlTemplateString, options) {
   return async state => {