+++ packages/fhir-4/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-fhir-4
 
+## 0.1.10 - 01 September 2025
+
+### Patch Changes
+
+- Updated dependencies \[1d60531]
+  - @openfn/language-common@3.0.3
+
 ## 0.1.9 - 14 July 2025
 
 ### Patch Changes