+++ packages/fhir-fr/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-fhir-fr
 
+## 1.0.16 - 01 September 2025
+
+### Patch Changes
+
+- Updated dependencies \[1d60531]
+  - @openfn/language-common@3.0.3
+
 ## 1.0.15 - 14 July 2025
 
 ### Patch Changes