+++ packages/fhir/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-fhir
 
+## 5.0.8 - 01 September 2025
+
+### Patch Changes
+
+- Updated dependencies \[1d60531]
+  - @openfn/language-common@3.0.3
+
 ## 5.0.7 - 14 July 2025
 
 ### Patch Changes