+++ packages/godata/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-godata
 
+## 3.5.8 - 01 September 2025
+
+### Patch Changes
+
+- Updated dependencies \[1d60531]
+  - @openfn/language-common@3.0.3
+
 ## 3.5.7 - 14 July 2025
 
 ### Patch Changes