+++ packages/hubtel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/language-hubtel
 
+## 1.0.11 - 01 September 2025
+
+### Patch Changes
+
+- Updated dependencies \[1d60531]
+  - @openfn/language-common@3.0.3
+
 ## 1.0.10 - 28 August 2025
 
 ### Patch Changes